// frontend/src/app/core/generated/circabc/model/userNewsFeed.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InterestGroupFeed } from './interestGroupFeed';


export interface UserNewsFeed { 
    when: UserNewsFeed.WhenEnum;
    uploads?: number;
    updates?: number;
    comments?: number;
    groupFeeds?: Array<InterestGroupFeed>;
}
export namespace UserNewsFeed {
    export const WhenEnum = {
        Today: 'today',
        Week: 'week',
        PreviousWeek: 'previousWeek'
    } as const;
    export type WhenEnum = typeof WhenEnum[keyof typeof WhenEnum];
}